migrated dose loading from BrachyPlan to catheter table
